+++ vtm-android-example/src/org/oscim/android/test/Samples.java
@@ -1,8 +1,9 @@
 /*
  * Copyright 2010, 2011, 2012, 2013 mapsforge.org
  * Copyright 2013 Hannes Janetzek
- * Copyright 2016 devemux86
+ * Copyright 2016-2017 devemux86
  * Copyright 2017 Longri
+ * Copyright 2017 nebular
  *
  * This file is part of the OpenScienceMap project (http://www.opensciencemap.org).
  *
@@ -53,6 +54,8 @@ public class Samples extends Activity {
         linearLayout.addView(createButton(MarkerOverlayActivity.class));
         linearLayout.addView(createButton(RotateMarkerOverlayActivity.class));
         linearLayout.addView(createButton(AtlasMarkerOverlayActivity.class));
+        linearLayout.addView(createButton(AtlasMultiTextureActivity.class));
+        linearLayout.addView(createButton(ClusterMarkerOverlayActivity.class));
         linearLayout.addView(createButton(PathOverlayActivity.class));
         linearLayout.addView(createButton(LineTexActivity.class));
         linearLayout.addView(createButton(VectorLayerMapActivity.class));