+++ vtm-android/src/org/oscim/android/canvas/AndroidCanvas.java
@@ -18,6 +18,7 @@
  */
 package org.oscim.android.canvas;
 
+import android.graphics.Color;
 import android.graphics.PorterDuff;
 
 import org.oscim.backend.canvas.Bitmap;
@@ -82,7 +83,7 @@ public class AndroidCanvas implements Canvas {
 
     @Override
     public void fillColor(int color) {
-        canvas.drawColor(color, PorterDuff.Mode.SRC);
+        canvas.drawColor(color, color == Color.TRANSPARENT ? PorterDuff.Mode.CLEAR : PorterDuff.Mode.SRC_OVER);
     }
 
     @Override